+++ ts/packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @composio/core
 
+## 0.3.3
+
+### Patch Changes
+
+- a76b002: Add support for enable/disable tags and search toolkits in tool router
+
+## 0.3.2
+
+### Patch Changes
+
+- 69cfede: Update client version and add openWorldHintSupport in toolrouter tag filters. Removes isLocal param in toolkit fetching
+
 ## 0.3.1
 
 ### Patch Changes