+++ ts/packages/providers/llamaindex/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @composio/llamaindex
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [a76b002]
+  - @composio/core@0.3.3
+
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [69cfede]
+  - @composio/core@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes